third_party/WebKit/LayoutTests/fast/css/pseudo-element-backdrop-hit-test.html

<!doctype html>
<script src="../../resources/js-test.js"></script>

<style>
#dialog::backdrop { background-color: green; }
</style>

<dialog id="target"></dialog>

<script>
description("Hit testing with ::backdrop style");

var dialog = document.getElementById("target");
dialog.showModal();

shouldBe("document.elementFromPoint(5, 5).id", "'target'");

dialog.close();
</script>
